Extract it_serverless fixture to conftest.py
